refactor(scripts): introduce barrel file to /prettier
