+++ scripts/prettier/index.d.ts
@@ -1,0 +1,1 @@
+export { runPrettierForFolder, runPrettier, prettierSupportedFileExtensionsByContext } from './index';
